requirements.txt
autopep8==1.7.0
keyboard==0.13.5
MouseInfo==0.1.3
PyAutoGUI==0.9.53
pycodestyle==2.9.1
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
toml==0.10.2
pywin32
numpy
Pillow
opencv-python